rename settings.js
