+++ src/defaultSettings.js
